#53 バランスを見て例え一覧ページの各要素の位置調整をした。
