+++ src/components/Layouts/DashBoard/SideBarLayouts.tsx
@@ -1,12 +1,12 @@
-import { MenuBtn } from '../../btn/MenuBtn';
+import { DashBoardMenuBtn } from '../../btn/DashBoardMenuBtn';
 import React, { VFC } from 'react';
 import { WithoutPropsChildrenLayouts } from '../../types/types';
 import { SideBarMainContentsLayouts } from './SideBarMainContentsLayouts';
 
-export const SideBarLayouts:VFC<WithoutPropsChildrenLayouts> = (props) => {
+export const SideBarLayouts: VFC<WithoutPropsChildrenLayouts> = (props) => {
   return (
     <aside
-    className='
+      className='
     bg-gray-900
     sm:min-w-[64px]
     sm:max-w-[64px]
@@ -19,15 +19,14 @@ export const SideBarLayouts:VFC<WithoutPropsChildrenLayouts> = (props) => {
     '
     >
       <nav
-      className='
+        className='
       flex
       md:flex-col
       flex-row'
       >
-        <MenuBtn />
+        <DashBoardMenuBtn />
         <SideBarMainContentsLayouts />
       </nav>
     </aside>
-
-  )
-}
+  );
+};
